+++ scripts/clean_up.sh
@@ -7,3 +7,6 @@ rm -rf build/
 rm -rf dist/
 rm -rf .cache/
 rm -rf htmlcov/
+rm -rf examples/django-rest-framework-tricks-env/
+rm -rf examples/rest_framework_tricks_demo_installer/
+rm examples/rest_framework_tricks_demo_installer.tar.gz