+++ src/rest_framework_tricks/__init__.py
@@ -5,7 +5,7 @@ Collection of various tricks for Django REST framework.
 __title__ = 'django-rest-framework-tricks'
 __version__ = '0.2.5'
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
-__copyright__ = '2017 Artur Barseghyan'
+__copyright__ = '2017-2018 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
 __all__ = ('default_app_config',)