+++ src/rest_framework_tricks/serializers/__init__.py
@@ -6,7 +6,7 @@ from .nested_proxy import *
 
 __title__ = 'rest_framework_tricks.serializers'
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
-__copyright__ = '2017 Artur Barseghyan'
+__copyright__ = '2017-2018 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
 __all__ = (
     'extract_nested_serializers',