+++ src/rest_framework_tricks/utils.py
@@ -7,7 +7,7 @@ from six import python_2_unicode_compatible
 
 __title__ = 'rest_framework_tricks.utils'
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
-__copyright__ = '2017 Artur Barseghyan'
+__copyright__ = '2017-2018 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
 __all__ = (
     'DictProxy',